Move function for scaling down to gcloud_update
