+++ scale/gcloud_update.py
@@ -43,6 +43,13 @@ def __get_singleuser_image(namespace, hub_pod, cluster_name):
 
     return image
 
+def shutdownSpecifiedNode(name):
+    """Deletes the specified node by calling the Google Cloud Engine"""
+    cmd = ['gcloud', 'compute', 'instance-groups', 'managed', 'delete-instances',
+           GCLOUD_INSTANCE_GROUP, '--instances=' + name]
+    p = subprocess.Popen(' '.join(cmd), stdout=subprocess.PIPE, shell=True)
+    output, err = p.communicate()
+    return "Successfully removed node: %s\n" % name if not err else "Unable to remove node"
 
 def increaseNewGCloudNode(new_node_number, cluster_name, namespaces):
     """ONLY FOR CREATING NEW NODES to ensure
