+++ scale/utils.py
@@ -122,14 +122,6 @@ def setUnschedulable(name, value=True, url=generateUrl(API_HOST, API_PORT) + "no
         except Exception:
             return "Return type was " + str(r.status)
 
-def shutdownSpecifiedNode(name):
-    """Deletes the specified node by calling the Google Cloud Engine"""
-    cmd = ['gcloud', 'compute', 'instance-groups', 'managed', 'delete-instances',
-	   GCLOUD_INSTANCE_GROUP, '--instances=' + name]
-    p = subprocess.Popen(' '.join(cmd), stdout=subprocess.PIPE, shell=True)
-    output, err = p.communicate()
-    return "Successfully removed node: %s\n" % name if not err else "Unable to remove node"
-    
 def getName(resource):
     """ Return name of a node, return '' if
     an error occurred """
